backlog-model/src/main/java/com/sitewidesystems/model/Story.java
package com.sitewidesystems.model;

import java.util.Date;
import java.util.List;

/**
 * A Story is the basic request component of the backlog system
 * User: gerwood
 * Date: Apr 5, 2009
 * Time: 2:11:16 PM
 */
public class Story {

    private String title;
    private String story;
    private String state;
    private String owner;
    private List<String> categories;
    private Date opened;
    private Date closed;
    private Number priority;
    private String project;


    public String getTitle() {
        return title;
    }

    public void setTitle(String title) {
        this.title = title;
    }

    public String getStory() {
        return story;
    }

    public void setStory(String story) {
        this.story = story;
    }

    public String getState() {
        return state;
    }

    public void setState(String state) {
        this.state = state;
    }

    public String getOwner() {
        return owner;
    }

    public void setOwner(String owner) {
        this.owner = owner;
    }

    public List<String> getCategories() {
        return categories;
    }

    public void setCategories(List<String> categories) {
        this.categories = categories;
    }

    public Date getOpened() {
        return opened;
    }

    public void setOpened(Date opened) {
        this.opened = opened;
    }

    public Number getPriority() {
        return priority;
    }

    public void setPriority(Number priority) {
        this.priority = priority;
    }

    public String getProject() {
        return project;
    }

    public void setProject(String project) {
        this.project = project;
    }

    public Date getClosed() {
        return closed;
    }

    public void setClosed(Date closed) {
        this.closed = closed;
    }
}
